+++ src/oci/generative_ai/models/create_dedicated_ai_cluster_details.py
@@ -229,6 +229,8 @@ class CreateDedicatedAiClusterDetails(object):
         - LARGE_GENERIC_4
         - SMALL_GENERIC_V2
         - LARGE_GENERIC_2
+        - LARGE_COHERE_V3
+        - RERANK_COHERE
 
 
         :return: The unit_shape of this CreateDedicatedAiClusterDetails.
@@ -255,6 +257,8 @@ class CreateDedicatedAiClusterDetails(object):
         - LARGE_GENERIC_4
         - SMALL_GENERIC_V2
         - LARGE_GENERIC_2
+        - LARGE_COHERE_V3
+        - RERANK_COHERE
 
 
         :param unit_shape: The unit_shape of this CreateDedicatedAiClusterDetails.