+++ src/oci/opensearch/models/resize_opensearch_cluster_vertical_details.py
@@ -30,6 +30,10 @@ class ResizeOpensearchClusterVerticalDetails(object):
             The value to assign to the master_node_host_memory_gb property of this ResizeOpensearchClusterVerticalDetails.
         :type master_node_host_memory_gb: int
 
+        :param master_node_host_shape:
+            The value to assign to the master_node_host_shape property of this ResizeOpensearchClusterVerticalDetails.
+        :type master_node_host_shape: str
+
         :param data_node_host_ocpu_count:
             The value to assign to the data_node_host_ocpu_count property of this ResizeOpensearchClusterVerticalDetails.
         :type data_node_host_ocpu_count: int
@@ -42,6 +46,10 @@ class ResizeOpensearchClusterVerticalDetails(object):
             The value to assign to the data_node_storage_gb property of this ResizeOpensearchClusterVerticalDetails.
         :type data_node_storage_gb: int
 
+        :param data_node_host_shape:
+            The value to assign to the data_node_host_shape property of this ResizeOpensearchClusterVerticalDetails.
+        :type data_node_host_shape: str
+
         :param opendashboard_node_host_ocpu_count:
             The value to assign to the opendashboard_node_host_ocpu_count property of this ResizeOpensearchClusterVerticalDetails.
         :type opendashboard_node_host_ocpu_count: int
@@ -49,6 +57,26 @@ class ResizeOpensearchClusterVerticalDetails(object):
         :param opendashboard_node_host_memory_gb:
             The value to assign to the opendashboard_node_host_memory_gb property of this ResizeOpensearchClusterVerticalDetails.
         :type opendashboard_node_host_memory_gb: int
+
+        :param opendashboard_node_host_shape:
+            The value to assign to the opendashboard_node_host_shape property of this ResizeOpensearchClusterVerticalDetails.
+        :type opendashboard_node_host_shape: str
+
+        :param search_node_host_shape:
+            The value to assign to the search_node_host_shape property of this ResizeOpensearchClusterVerticalDetails.
+        :type search_node_host_shape: str
+
+        :param search_node_host_ocpu_count:
+            The value to assign to the search_node_host_ocpu_count property of this ResizeOpensearchClusterVerticalDetails.
+        :type search_node_host_ocpu_count: int
+
+        :param search_node_host_memory_gb:
+            The value to assign to the search_node_host_memory_gb property of this ResizeOpensearchClusterVerticalDetails.
+        :type search_node_host_memory_gb: int
+
+        :param search_node_storage_gb:
+            The value to assign to the search_node_storage_gb property of this ResizeOpensearchClusterVerticalDetails.
+        :type search_node_storage_gb: int
 
         :param freeform_tags:
             The value to assign to the freeform_tags property of this ResizeOpensearchClusterVerticalDetails.
@@ -62,32 +90,53 @@ class ResizeOpensearchClusterVerticalDetails(object):
         self.swagger_types = {
             'master_node_host_ocpu_count': 'int',
             'master_node_host_memory_gb': 'int',
+            'master_node_host_shape': 'str',
             'data_node_host_ocpu_count': 'int',
             'data_node_host_memory_gb': 'int',
             'data_node_storage_gb': 'int',
+            'data_node_host_shape': 'str',
             'opendashboard_node_host_ocpu_count': 'int',
             'opendashboard_node_host_memory_gb': 'int',
+            'opendashboard_node_host_shape': 'str',
+            'search_node_host_shape': 'str',
+            'search_node_host_ocpu_count': 'int',
+            'search_node_host_memory_gb': 'int',
+            'search_node_storage_gb': 'int',
             'freeform_tags': 'dict(str, str)',
             'defined_tags': 'dict(str, dict(str, object))'
         }
         self.attribute_map = {
             'master_node_host_ocpu_count': 'masterNodeHostOcpuCount',
             'master_node_host_memory_gb': 'masterNodeHostMemoryGB',
+            'master_node_host_shape': 'masterNodeHostShape',
             'data_node_host_ocpu_count': 'dataNodeHostOcpuCount',
             'data_node_host_memory_gb': 'dataNodeHostMemoryGB',
             'data_node_storage_gb': 'dataNodeStorageGB',
+            'data_node_host_shape': 'dataNodeHostShape',
             'opendashboard_node_host_ocpu_count': 'opendashboardNodeHostOcpuCount',
             'opendashboard_node_host_memory_gb': 'opendashboardNodeHostMemoryGB',
+            'opendashboard_node_host_shape': 'opendashboardNodeHostShape',
+            'search_node_host_shape': 'searchNodeHostShape',
+            'search_node_host_ocpu_count': 'searchNodeHostOcpuCount',
+            'search_node_host_memory_gb': 'searchNodeHostMemoryGB',
+            'search_node_storage_gb': 'searchNodeStorageGB',
             'freeform_tags': 'freeformTags',
             'defined_tags': 'definedTags'
         }
         self._master_node_host_ocpu_count = None
         self._master_node_host_memory_gb = None
+        self._master_node_host_shape = None
         self._data_node_host_ocpu_count = None
         self._data_node_host_memory_gb = None
         self._data_node_storage_gb = None
+        self._data_node_host_shape = None
         self._opendashboard_node_host_ocpu_count = None
         self._opendashboard_node_host_memory_gb = None
+        self._opendashboard_node_host_shape = None
+        self._search_node_host_shape = None
+        self._search_node_host_ocpu_count = None
+        self._search_node_host_memory_gb = None
+        self._search_node_storage_gb = None
         self._freeform_tags = None
         self._defined_tags = None
 
@@ -138,6 +187,30 @@ class ResizeOpensearchClusterVerticalDetails(object):
         :type: int
         """
         self._master_node_host_memory_gb = master_node_host_memory_gb
+
+    @property
+    def master_node_host_shape(self):
+        """
+        Gets the master_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's master nodes.
+
+
+        :return: The master_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: str
+        """
+        return self._master_node_host_shape
+
+    @master_node_host_shape.setter
+    def master_node_host_shape(self, master_node_host_shape):
+        """
+        Sets the master_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's master nodes.
+
+
+        :param master_node_host_shape: The master_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :type: str
+        """
+        self._master_node_host_shape = master_node_host_shape
 
     @property
     def data_node_host_ocpu_count(self):
@@ -212,6 +285,30 @@ class ResizeOpensearchClusterVerticalDetails(object):
         self._data_node_storage_gb = data_node_storage_gb
 
     @property
+    def data_node_host_shape(self):
+        """
+        Gets the data_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's data nodes.
+
+
+        :return: The data_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: str
+        """
+        return self._data_node_host_shape
+
+    @data_node_host_shape.setter
+    def data_node_host_shape(self, data_node_host_shape):
+        """
+        Sets the data_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's data nodes.
+
+
+        :param data_node_host_shape: The data_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :type: str
+        """
+        self._data_node_host_shape = data_node_host_shape
+
+    @property
     def opendashboard_node_host_ocpu_count(self):
         """
         Gets the opendashboard_node_host_ocpu_count of this ResizeOpensearchClusterVerticalDetails.
@@ -258,6 +355,126 @@ class ResizeOpensearchClusterVerticalDetails(object):
         :type: int
         """
         self._opendashboard_node_host_memory_gb = opendashboard_node_host_memory_gb
+
+    @property
+    def opendashboard_node_host_shape(self):
+        """
+        Gets the opendashboard_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's OpenSearch Dashboard nodes.
+
+
+        :return: The opendashboard_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: str
+        """
+        return self._opendashboard_node_host_shape
+
+    @opendashboard_node_host_shape.setter
+    def opendashboard_node_host_shape(self, opendashboard_node_host_shape):
+        """
+        Sets the opendashboard_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's OpenSearch Dashboard nodes.
+
+
+        :param opendashboard_node_host_shape: The opendashboard_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :type: str
+        """
+        self._opendashboard_node_host_shape = opendashboard_node_host_shape
+
+    @property
+    def search_node_host_shape(self):
+        """
+        Gets the search_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's search nodes.
+
+
+        :return: The search_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: str
+        """
+        return self._search_node_host_shape
+
+    @search_node_host_shape.setter
+    def search_node_host_shape(self, search_node_host_shape):
+        """
+        Sets the search_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        The node shape for the cluster's search nodes.
+
+
+        :param search_node_host_shape: The search_node_host_shape of this ResizeOpensearchClusterVerticalDetails.
+        :type: str
+        """
+        self._search_node_host_shape = search_node_host_shape
+
+    @property
+    def search_node_host_ocpu_count(self):
+        """
+        Gets the search_node_host_ocpu_count of this ResizeOpensearchClusterVerticalDetails.
+        The number of OCPUs configured for the cluster's search nodes.
+
+
+        :return: The search_node_host_ocpu_count of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: int
+        """
+        return self._search_node_host_ocpu_count
+
+    @search_node_host_ocpu_count.setter
+    def search_node_host_ocpu_count(self, search_node_host_ocpu_count):
+        """
+        Sets the search_node_host_ocpu_count of this ResizeOpensearchClusterVerticalDetails.
+        The number of OCPUs configured for the cluster's search nodes.
+
+
+        :param search_node_host_ocpu_count: The search_node_host_ocpu_count of this ResizeOpensearchClusterVerticalDetails.
+        :type: int
+        """
+        self._search_node_host_ocpu_count = search_node_host_ocpu_count
+
+    @property
+    def search_node_host_memory_gb(self):
+        """
+        Gets the search_node_host_memory_gb of this ResizeOpensearchClusterVerticalDetails.
+        The amount of memory in GB, for the cluster's search nodes.
+
+
+        :return: The search_node_host_memory_gb of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: int
+        """
+        return self._search_node_host_memory_gb
+
+    @search_node_host_memory_gb.setter
+    def search_node_host_memory_gb(self, search_node_host_memory_gb):
+        """
+        Sets the search_node_host_memory_gb of this ResizeOpensearchClusterVerticalDetails.
+        The amount of memory in GB, for the cluster's search nodes.
+
+
+        :param search_node_host_memory_gb: The search_node_host_memory_gb of this ResizeOpensearchClusterVerticalDetails.
+        :type: int
+        """
+        self._search_node_host_memory_gb = search_node_host_memory_gb
+
+    @property
+    def search_node_storage_gb(self):
+        """
+        Gets the search_node_storage_gb of this ResizeOpensearchClusterVerticalDetails.
+        The amount of storage in GB, to configure per node for the cluster's search nodes.
+
+
+        :return: The search_node_storage_gb of this ResizeOpensearchClusterVerticalDetails.
+        :rtype: int
+        """
+        return self._search_node_storage_gb
+
+    @search_node_storage_gb.setter
+    def search_node_storage_gb(self, search_node_storage_gb):
+        """
+        Sets the search_node_storage_gb of this ResizeOpensearchClusterVerticalDetails.
+        The amount of storage in GB, to configure per node for the cluster's search nodes.
+
+
+        :param search_node_storage_gb: The search_node_storage_gb of this ResizeOpensearchClusterVerticalDetails.
+        :type: int
+        """
+        self._search_node_storage_gb = search_node_storage_gb
 
     @property
     def freeform_tags(self):